Reorganize file structure, fix warnings, remove unused code, and run clang format.
