+++ VCTwoFluidStaggeredStokesOperator.cpp
@@ -39,15 +39,13 @@
 #include "tbox/TimerManager.h"
 
 #include <algorithm>
+#include <cmath>
 #include <string>
 #include <vector>
 
 // Local includes
 #include "VCTwoFluidStaggeredStokesOperator.h"
 
-#include <math.h>
-
-#include <string>
 /////////////////////////////// NAMESPACE ////////////////////////////////////
 
 namespace IBAMR
@@ -65,9 +63,10 @@ static const int SIDEG = 1;
 static const std::string CC_DATA_REFINE_TYPE =
     "CONSERVATIVE_LINEAR_REFINE"; // how to fill in fine cells from coarse cells, how to fill ghost cells on refine
                                   // patch
-static const std::string SC_DATA_REFINE_TYPE = "CONSERVATIVE_LINEAR_REFINE"; // how to fill in fine cells from coarse cells, how to fill ghost
-                                                       // cells on refine patch
-static const bool USE_CF_INTERPOLATION = true; // Refine Patch Strategy: CartSideDoubleQuadraticCFInterpolation. 
+static const std::string SC_DATA_REFINE_TYPE =
+    "CONSERVATIVE_LINEAR_REFINE";              // how to fill in fine cells from coarse cells, how to fill ghost
+                                               // cells on refine patch
+static const bool USE_CF_INTERPOLATION = true; // Refine Patch Strategy: CartSideDoubleQuadraticCFInterpolation.
 static const std::string DATA_COARSEN_TYPE =
     "CUBIC_COARSEN"; // going from fine to coarse. fill in coarse cells by whatever is in the fine cells. synchronizing
                      // the hierarchies
@@ -83,65 +82,13 @@ static const bool CONSISTENT_TYPE_2_BDRY = false;
 static Timer* t_apply;
 static Timer* t_initialize_operator_state;
 static Timer* t_deallocate_operator_state;
-
-/// Functions with exact solutions for error checking.
-VectorNd
-un_fcn(const VectorNd& x)
-{
-    VectorNd vel;
-    vel(0) = std::cos(2.0 * M_PI * x(0)) * std::cos(2.0 * M_PI * x(1));
-    vel(1) = std::cos(2.0 * M_PI * x(1)) * std::cos(2.0 * M_PI * x(0));
-    return vel;
-}
-VectorNd
-us_fcn(const VectorNd& x)
-{
-    VectorNd vel;
-    vel(0) = -std::cos(2.0 * M_PI * x(0)) * std::cos(2.0 * M_PI * x(1));
-    vel(1) = -std::cos(2.0 * M_PI * x(1)) * std::cos(2.0 * M_PI * x(0));
-    return vel;
-}
-double
-thetan(const VectorNd& x)
-{
-    return 0.25 * std::sin(2.0 * M_PI * x(0)) * std::sin(2.0 * M_PI * x(1)) + 0.5;
-}
-
-double
-p_fcn(const VectorNd& x)
-{
-    return 0.0;
-}
-
-double
-dy_ths_dus_dy(const VectorNd& x)
-{
-    return -2.0 * M_PI * M_PI * std::cos(2.0 * M_PI * x(0)) * std::cos(2.0 * M_PI * x[1]) *
-           (-1.0 + std::sin(2.0 * M_PI * x(0)) * std::sin(2.0 * M_PI * x(1)));
-}
-double
-dx_ths_dus_dx(const VectorNd& x)
-{
-    return -2.0 * M_PI * M_PI * std::cos(2.0 * M_PI * x(0)) * std::cos(2.0 * M_PI * x(1)) *
-           (-1.0 + std::sin(2.0 * M_PI * x(0)) * std::sin(2.0 * M_PI * x(1)));
-}
-double
-dx_ths_dus_dy(const VectorNd& x)
-{
-    return -0.5 * M_PI * M_PI * std::sin(2.0 * M_PI * x(1)) *
-           (4.0 * std::sin(2.0 * M_PI * x(0)) + std::sin(2.0 * M_PI * (2.0 * x(0) + x(1))) -
-            std::sin(2.0 * M_PI * (2.0 * x(0) - x(1))));
-}
-double
-dy_ths_dus_dx(const VectorNd& x)
-{
-    return -0.5 * M_PI * M_PI * std::sin(2.0 * M_PI * x(0)) *
-           (4.0 * std::sin(2.0 * M_PI * x(1)) + std::sin(2.0 * M_PI * (x(0) - 2.0 * x(1))) +
-            std::sin(2.0 * M_PI * (x(0) + 2.0 * x(1))));
-}
 } // namespace
 
-double convertToThs(double Thn);
+double
+convertToThs(double Thn)
+{
+    return 1.0 - Thn; // Thn+Ths = 1
+}
 /////////////////////////////// PUBLIC ///////////////////////////////////////
 VCTwoFluidStaggeredStokesOperator::VCTwoFluidStaggeredStokesOperator(const std::string& object_name,
                                                                      bool homogeneous_bc,
@@ -374,9 +321,6 @@ VCTwoFluidStaggeredStokesOperator::apply(SAMRAIVectorReal<NDIM, double>& x, SAMR
             Pointer<Patch<NDIM>> patch = level->getPatch(p());
             Pointer<CartesianPatchGeometry<NDIM>> pgeom = patch->getPatchGeometry();
             const double* const dx = pgeom->getDx(); // dx[0] -> x, dx[1] -> y
-            const double* const xlow =
-                pgeom->getXLower(); // {xlow[0], xlow[1]} -> physical location of bottom left of box.
-            const hier::Index<NDIM>& idx_low = patch->getBox().lower();
             Pointer<CellData<NDIM, double>> p_data = patch->getPatchData(P_idx);
             Pointer<CellData<NDIM, double>> A_P_data =
                 patch->getPatchData(A_P_idx); // result of applying operator (eqn 3)
@@ -389,48 +333,9 @@ VCTwoFluidStaggeredStokesOperator::apply(SAMRAIVectorReal<NDIM, double>& x, SAMR
                 patch->getPatchData(A_us_idx); // result of applying operator (eqn 2)
             IntVector<NDIM> xp(1, 0), yp(0, 1);
 
-            // Fill ghost cells by hand
-#if (0)
-            for (int axis = 0; axis < NDIM; ++axis)
-            {
-                for (SideIterator<NDIM> si(us_data->getGhostBox(), axis); si; si++)
-                {
-                    const SideIndex<NDIM>& idx = si();
-                    if (!SideGeometry<NDIM>::toSideBox(patch->getBox(), axis).contains(idx))
-                    {
-                        VectorNd x;
-                        for (int d = 0; d < NDIM; ++d)
-                            x[d] =
-                                xlow[d] + dx[d] * (static_cast<double>(idx(d) - idx_low(d)) + (d == axis ? 0.0 : 0.5));
-
-                        VectorNd un_vel = un_fcn(x);
-                        (*un_data)(idx) = un_vel(axis);
-                        VectorNd us_vel = us_fcn(x);
-                        (*us_data)(idx) = us_vel(axis);
-                    }
-                }
-            }
-            for (CellIterator<NDIM> ci(thn_data->getGhostBox()); ci; ci++)
-            {
-                const CellIndex<NDIM>& idx = ci();
-                if (!patch->getBox().contains(idx))
-                {
-                    VectorNd x;
-                    for (int d = 0; d < NDIM; ++d)
-                        x[d] = xlow[d] + dx[d] * (static_cast<double>(idx(d) - idx_low(d)) + 0.5);
-                    (*thn_data)(idx) = thetan(x);
-                    (*p_data)(idx) = p_fcn(x);
-                }
-            }
-#endif
-
             for (CellIterator<NDIM> ci(patch->getBox()); ci; ci++) // cell-centers
             {
                 const CellIndex<NDIM>& idx = ci();
-                VectorNd x; // <- Eigen3 vector
-                for (int d = 0; d < NDIM; ++d)
-                    x[d] = xlow[d] + dx[d] * (idx(d) - idx_low(d) + 0.5); // Get's physical location of idx.
-
                 SideIndex<NDIM> lower_x_idx(idx, 0, 0); // (i-1/2,j)
                 SideIndex<NDIM> upper_x_idx(idx, 0, 1); // (i+1/2,j)
                 SideIndex<NDIM> lower_y_idx(idx, 1, 0); // (i,j-1/2)
@@ -596,7 +501,8 @@ VCTwoFluidStaggeredStokesOperator::apply(SAMRAIVectorReal<NDIM, double>& x, SAMR
     }
     if (d_bc_helper) d_bc_helper->copyDataAtDirichletBoundaries(A_un_idx, un_scratch_idx);
 
-    auto sync_fcn = [&](const int dst_idx) -> void {
+    auto sync_fcn = [&](const int dst_idx) -> void
+    {
         for (int ln = d_hierarchy->getFinestLevelNumber(); ln > 0; --ln)
         {
             Pointer<PatchLevel<NDIM>> level = d_hierarchy->getPatchLevel(ln);
@@ -611,7 +517,7 @@ VCTwoFluidStaggeredStokesOperator::apply(SAMRAIVectorReal<NDIM, double>& x, SAMR
             coarsen_alg->registerCoarsen(dst_idx, d_os_idx, d_os_coarsen_op);
             coarsen_alg->resetSchedule(d_os_coarsen_scheds[ln - 1]);
             d_os_coarsen_scheds[ln - 1]->coarsenData();
-            d_os_coarsen_alg->resetSchedule(d_os_coarsen_scheds[ln - 1]); 
+            d_os_coarsen_alg->resetSchedule(d_os_coarsen_scheds[ln - 1]);
         }
     };
 
@@ -621,12 +527,6 @@ VCTwoFluidStaggeredStokesOperator::apply(SAMRAIVectorReal<NDIM, double>& x, SAMR
     IBAMR_TIMER_STOP(t_apply);
     return;
 } // apply
-
-double
-convertToThs(double Thn)
-{
-    return 1.0 - Thn; // Thn+Ths = 1
-}
 
 void
 VCTwoFluidStaggeredStokesOperator::initializeOperatorState(const SAMRAIVectorReal<NDIM, double>& in,
