+++ smooth.cpp
@@ -272,10 +272,8 @@ main(int argc, char* argv[])
         p_fcn.setDataOnPatchHierarchy(p_cc_idx, p_cc_var, patch_hierarchy, 0.0);
 
         // Setup the box relaxation FAC operator
-
-        VCTwoFluidStaggeredStokesBoxRelaxationFACOperator box_relax("box_relax", "");
+        VCTwoFluidStaggeredStokesBoxRelaxationFACOperator box_relax("box_relax", "", 1.0 /*w*/, 0.0 /*C*/, 1.0 /*D*/);
         box_relax.setThnIdx(thn_cc_idx);
-        // box_relax.initializeOperatorState(u_vec,f_vec);
         box_relax.setToZero(u_vec, 0);
 
         for (int i = 0; i <= 680; i++)
@@ -285,52 +283,6 @@ main(int argc, char* argv[])
             pout << "Sweep = " << i << "\n";
             pout << "|r|_2  = " << e_vec.L2Norm() << "\n";
         }
-
-        // Setup the stokes operator
-        // VCTwoFluidStaggeredStokesOperator stokes_op("stokes_op", true);
-        // stokes_op.setThnIdx(thn_cc_idx);
-        // stokes_op.initializeOperatorState(u_vec, e_vec);
-        // stokes_op.apply(u_vec, e_vec); // e_vec is A*u
-
-        // calculate residual norm: b - A*u
-        // f_vec.subtract(Pointer<SAMRAIVectorReal<NDIM, double>>(&f_vec, false),  // RHS
-        // Pointer<SAMRAIVectorReal<NDIM, double>>(&e_vec, false)); // A*u
-        // pout << "|r|_oo = " << e_vec.maxNorm() << "\n";
-        // pout << "|r|_2  = " << e_vec.L2Norm() << "\n";
-        // pout << "|r|_1  = " << e_vec.L1Norm() << "\n";
-
-        // Compute error and print error norms.
-        // e_vec.subtract(Pointer<SAMRAIVectorReal<NDIM, double>>(&u_vec, false),  // numerical
-        //                Pointer<SAMRAIVectorReal<NDIM, double>>(&e_vec, false)); // analytical
-        // pout << "|e|_oo = " << e_vec.maxNorm() << "\n";
-        // pout << "|e|_2  = " << e_vec.L2Norm() << "\n";
-        // pout << "|e|_1  = " << e_vec.L1Norm() << "\n";
-
-        // hier_math_ops.setPatchHierarchy(patch_hierarchy);
-        // hier_math_ops.resetLevels(0, patch_hierarchy->getFinestLevelNumber());
-        // // just computes quadrature weights
-        // const int wgt_cc_idx = hier_math_ops.getCellWeightPatchDescriptorIndex();
-        // const int wgt_sc_idx = hier_math_ops.getSideWeightPatchDescriptorIndex();
-
-        // HierarchySideDataOpsReal<NDIM, double> hier_sc_data_ops(
-        //     patch_hierarchy, 0, patch_hierarchy->getFinestLevelNumber());
-        // pout << "Error in u_n :\n"
-        //      << "  L1-norm:  " << std::setprecision(10) << hier_sc_data_ops.L1Norm(e_un_sc_idx, wgt_sc_idx) << "\n"
-        //      << "  L2-norm:  " << hier_sc_data_ops.L2Norm(e_un_sc_idx, wgt_sc_idx) << "\n"
-        //      << "  max-norm: " << hier_sc_data_ops.maxNorm(e_un_sc_idx, wgt_sc_idx) << "\n";
-
-        // pout << "Error in u_s :\n"
-        //      << "  L1-norm:  " << std::setprecision(10) << hier_sc_data_ops.L1Norm(e_us_sc_idx, wgt_sc_idx) << "\n"
-        //      << "  L2-norm:  " << hier_sc_data_ops.L2Norm(e_us_sc_idx, wgt_sc_idx) << "\n"
-        //      << "  max-norm: " << hier_sc_data_ops.maxNorm(e_us_sc_idx, wgt_sc_idx) << "\n";
-
-        // HierarchyCellDataOpsReal<NDIM, double> hier_cc_data_ops(
-        //     patch_hierarchy, 0, patch_hierarchy->getFinestLevelNumber());
-        // pout << "Error in p :\n"
-        //      << "  L1-norm:  " << hier_cc_data_ops.L1Norm(e_cc_idx, wgt_cc_idx) << "\n"
-        //      << "  L2-norm:  " << hier_cc_data_ops.L2Norm(e_cc_idx, wgt_cc_idx) << "\n"
-        //      << "  max-norm: " << hier_cc_data_ops.maxNorm(e_cc_idx, wgt_cc_idx) << "\n"
-        //      << "+++++++++++++++++++++++++++++++++++++++++++++++++++\n";
 
         // Interpolate the side-centered data to cell centers for output.
         static const bool synch_cf_interface = true;
